merge

+++ dal-dao-gen/src/main/java/com/ctrip/platform/dal/daogen/AbstractGenerator.java
@@ -1,10 +1,17 @@
 package com.ctrip.platform.dal.daogen;
 
 import java.io.File;
+import java.io.FileWriter;
 import java.io.IOException;
+import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 import org.apache.commons.io.FileUtils;
+import org.apache.commons.lang.StringUtils;
+import org.apache.commons.lang.WordUtils;
+import org.apache.velocity.VelocityContext;
 import org.apache.velocity.app.Velocity;
 
 import com.ctrip.platform.dal.daogen.dao.DaoByFreeSql;
@@ -12,10 +19,12 @@ import com.ctrip.platform.dal.daogen.dao.DaoBySqlBuilder;
 import com.ctrip.platform.dal.daogen.dao.DaoByTableViewSp;
 import com.ctrip.platform.dal.daogen.dao.DaoOfDbServer;
 import com.ctrip.platform.dal.daogen.dao.DaoOfProject;
+import com.ctrip.platform.dal.daogen.pojo.DbServer;
 import com.ctrip.platform.dal.daogen.pojo.GenTask;
 import com.ctrip.platform.dal.daogen.pojo.GenTaskByFreeSql;
 import com.ctrip.platform.dal.daogen.pojo.GenTaskByTableViewSp;
 import com.ctrip.platform.dal.daogen.pojo.Project;
+import com.ctrip.platform.dal.daogen.utils.JavaIOUtils;
 import com.ctrip.platform.dal.daogen.utils.SpringBeanGetter;
 
 public abstract class AbstractGenerator implements Generator {
@@ -64,7 +73,8 @@ public abstract class AbstractGenerator implements Generator {
 				projectId));
 		
 		try {
-			FileUtils.forceDelete(mavenLikeDir);
+			if(mavenLikeDir.exists())
+				FileUtils.forceDelete(mavenLikeDir);
 			//FileUtils.forceMkdir(mavenLikeDir);
 			File daoMavenLike = new File(mavenLikeDir, "Dao");
 			File entityMavenLike = new File(mavenLikeDir, "Entity");
@@ -76,8 +86,80 @@ public abstract class AbstractGenerator implements Generator {
 		} catch (IOException e1) {
 			e1.printStackTrace();
 		}
+		
+		Map<String,String> dbs = new HashMap<String, String>();
+		List<String> clazz = new ArrayList<String>();
+		
+		List<GenTaskByFreeSql> freeSql = daoByFreeSql.getTasksByProjectId(projectId);
+		List<GenTaskByTableViewSp> tableViewSps = daoByTableViewSp.getTasksByProjectId(projectId);
+		
+		for(GenTaskByFreeSql task : freeSql){
+			if(!dbs.containsKey(task.getDb_name())){
+				DbServer dbServer = daoOfDbServer.getDbServerByID(task.getServer_id());
+				String provider = "sqlProvider";
+				if (dbServer.getDb_type().equalsIgnoreCase("mysql")) {
+					provider = "mySqlProvider";
+				}
+				dbs.put(task.getDb_name(), provider);
+			}
+		}
+		for(GenTaskByTableViewSp task : tableViewSps){
+			if(!dbs.containsKey(task.getDb_name())){
+				DbServer dbServer = daoOfDbServer.getDbServerByID(task.getServer_id());
+				String provider = "sqlProvider";
+				if (dbServer.getDb_type().equalsIgnoreCase("mysql")) {
+					provider = "mySqlProvider";
+				}
+				dbs.put(task.getDb_name(), provider);
+			}
+			for(String t : StringUtils.split(task.getTable_names(), ",")){
+				if(!clazz.contains(t)){
+					String className = t;
+					if (null != task.getPrefix() && !task.getPrefix().isEmpty()) {
+						className = className.substring(task.getPrefix().length());
+					}
+					if (null != task.getSuffix() && !task.getSuffix() .isEmpty()) {
+						className = className + task.getSuffix() ;
+					}
+					clazz.add(className);
+				}
+			}
+		}
+		
+		VelocityContext context = new VelocityContext();
+		
+		context.put("dbs", dbs);
+		context.put("clazzList", clazz);
+		context.put("namespace", namespace);
+		context.put("WordUtils", WordUtils.class);
+		
+		FileWriter dalConfigWriter = null;
+		FileWriter dalFactoryWriter = null;
+		FileWriter dalProgramWriter = null;
+		try {
+		
+			dalConfigWriter = new FileWriter(String.format("%s/Dal.config",
+					mavenLikeDir.getAbsolutePath()));
+			dalFactoryWriter = new FileWriter(String.format("%s/DalFactory.cs",
+					mavenLikeDir.getAbsolutePath()));
+			dalProgramWriter = new FileWriter(String.format("%s/Program.cs",
+					mavenLikeDir.getAbsolutePath()));
 
-		generateByTableView(daoByTableViewSp.getTasksByProjectId(projectId));
+			Velocity.mergeTemplate("templates/Dal.config.tpl", "UTF-8",
+					context, dalConfigWriter);
+			Velocity.mergeTemplate("templates/DalFactory.cs.tpl", "UTF-8",
+					context, dalFactoryWriter);
+			Velocity.mergeTemplate("templates/Program.cs.tpl", "UTF-8",
+					context, dalProgramWriter);
+		} catch (IOException e) {
+			e.printStackTrace();
+		} finally {
+			JavaIOUtils.closeWriter(dalConfigWriter);
+			JavaIOUtils.closeWriter(dalFactoryWriter);
+			JavaIOUtils.closeWriter(dalProgramWriter);
+		}
+
+		generateByTableView(tableViewSps);
 
 //		// 存储过程
 //		List<GenTaskBySP> sp = spTaskDao.getTasksByProjectId(Integer
@@ -89,7 +171,7 @@ public abstract class AbstractGenerator implements Generator {
 //		generateBySP(_sp);
 //
 		// 手工编写的SQL
-		List<GenTaskByFreeSql> freeSql = daoByFreeSql.getTasksByProjectId(projectId);
+		
 
 		generateByFreeSql(freeSql);
